requirements.txt
django
djangorestframework

mysqlclient

coverage
